Working bounding boxes
The bounding boxes are more like circles, but it works better for
sprites with their origin at the center. Since it is not important that
the actual bounds of the sprite collide, this will be used to implement
an if-else statement telling the program whether or not it needs to
activate collision detection in the future.